requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==24.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Bottleneck==1.4.2
# Brotli==1.0.9
cachetools==5.5.1
camel-converter==4.0.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
curl_cffi==0.11.1
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
executing==2.1.0
frozendict==2.4.6
gitdb==4.0.7
GitPython==3.1.43
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.25
langchain-core==0.3.61
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
meilisearch==0.34.1
# mkl_fft==1.3.14
# mkl_random==1.2.8
# mkl-service==2.4.0
multitasking==0.0.11
narwhals==1.31.0
nest-asyncio==1.6.0
numexpr==2.10.1
numpy==2.2.5
openai==1.82.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
parso==0.8.4
peewee==3.18.1
pillow==11.1.0
pip==25.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.1
pure_eval==0.2.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
# pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.30.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.22.3
setuptools==72.1.0
six==1.17.0
smmap==4.0.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
streamlit==1.45.1
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.3.0
watchdog==4.0.2
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
win_inet_pton==1.1.0
yfinance==0.2.61
zstandard==0.23.0
